pyproject.toml
[project]
name = "propagate"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = ["pyzmq>=26.0", "pyyaml>=6.0", "python-dotenv>=1.0"]

[project.optional-dependencies]
webhook = ["fastapi>=0.100", "uvicorn[standard]>=0.20"]
telegram = ["python-telegram-bot>=21.0"]
dev = ["pytest", "httpx", "anyio", "pytest-anyio"]

[project.scripts]
propagate = "propagate_app.cli:main"
propagate-webhook = "propagate_webhook.cli:main"
propagate-telegram = "propagate_telegram.cli:main"

[tool.setuptools.packages.find]
include = ["propagate_app*", "propagate_webhook*", "propagate_telegram*"]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
markers = ["slow: tests that take >1s (git, serve, zmq)"]

[tool.ruff]
target-version = "py310"
line-length = 140

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]
ignore = ["E501"]
